Moved out character enum functions
