+++ src/Engine/PriceCalculator.cpp
@@ -5,6 +5,7 @@
 #include "Engine/Tables/BuildingTable.h"
 #include "Engine/Objects/Items.h"
 #include "Engine/Objects/Character.h"
+#include "Engine/Objects/CharacterEnumFunctions.h"
 #include "Engine/Party.h"
 
 int PriceCalculator::baseItemIdentifyPrice(float priceMultiplier) {
