+++ src/GUI/UI/UIQuickReference.cpp
@@ -2,7 +2,7 @@
 
 #include "Engine/AssetsManager.h"
 #include "Engine/Engine.h"
-#include "Engine/EngineGlobals.h"
+#include "Engine/Objects/CharacterEnumFunctions.h"
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Spells/Spells.h"
 #include "Engine/Localization.h"
@@ -11,7 +11,6 @@
 
 #include "GUI/GUIButton.h"
 #include "GUI/GUIFont.h"
-#include "GUI/UI/UICharacter.h"
 #include "GUI/UI/UIQuickReference.h"
 
 #include "Media/Audio/AudioPlayer.h"
