+++ src/GUI/UI/UISpellbook.cpp
@@ -5,6 +5,7 @@
 #include "Engine/AssetsManager.h"
 #include "Engine/Engine.h"
 #include "Engine/EngineGlobals.h"
+#include "Engine/Objects/CharacterEnumFunctions.h"
 #include "Engine/Graphics/IRender.h"
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/Graphics/Image.h"
@@ -14,7 +15,6 @@
 #include "Engine/Time.h"
 
 #include "GUI/GUIButton.h"
-#include "GUI/GUIFont.h"
 
 #include "Io/Mouse.h"
